+++ src/App.js
@@ -1,15 +1,12 @@
 import React from 'react';
 import './App.css';
-import Login from './assets/components/login/Login';
-import { useUser } from 'reactfire';
+import { LoginScreen } from './assets/components/login/LoginScreen';
 
 function App() {
-  const user = useUser();
 
   return (
     <div className="App">
-      { user && <p>Bienvenidx: {user.email} </p> }
-      <Login />
+      <LoginScreen />
     </div>
   );
 }
