+++ src/assets/components/login/Login.css
@@ -1,29 +1,96 @@
+.login-screen {
+    display: flex;
+}
 
-.btnLogin = {
-    background: #525252;
-    border: 0;
-    borderRadius: 25px;
-    color: #525252;
-    fontFamily: Ubuntu;
-    fontSize: 18px;
-    fontWeight: bold;
-    marginTop: 15%;
-    padding: 8px 0;
-    textAlign: center;
-    width: 180px;
+.login-box-left {
+    width: 35%;
+}
+
+.login-logo {
+    margin-top: 20%;
+}
+
+.login-h1 {
+    color: #8CAFAE;
+    font-size: 30px;
+    margin-left: 95px;
+    text-align: left;
+    width: 277px;
+}
+
+.login-h3 {
+    color: #2A303A;
+    font-size: 21px;
+    margin-left: 95px;
+    text-align: left;
+    width: 277px;
+}
+
+h5 {
+    color: #2A303A;
+    font-size: 16px;
+    margin-left: 95px;
+    text-align: left;
+    width: 277px;
+}
+
+.input-login {
+    background: #FFFFFF;
+    border: 1px solid #9DC5BB;
+    border-radius: 5px;
+    margin-left: 95px;
+    padding: 10px;
+    width: 260px;
+
+    font-family: Mada;
+    font-size: 12px;
+    line-height: 15px;
+    display: flex;
+    color: #ACB3B3;
+}
+
+.section-6, .section-7 {
+    display: flex;
+    width: 277px;
+    margin-left: 95px;
+}
+
+.h6 {
+    color: #2A303A;
+    font-size: 12px;
+    font-weight: bold;
+    text-align: left;
+    margin-right: auto;
+}
+
+.link-login {
+    color: #C78D00;
+}
+
+.btn-login {
+    background: #00BCA0;
+    border: none;
+    border-radius: 5px;
+    padding: 10px;
+    margin-left: 95px;
+    width: 265px;
+
+    font-family: Mada;
+    font-size: 16px;
+    text-align: center;
+    color: #FFFFFF;
 }
 
 
-.inputLogin = {
-    width: 350px;
-    height: 30px;
-    marginTop : 4%;
-    borderRadius: 15px;
-    background: #FFFFFF;
-    fontFamily: Roboto;
-    fontStyle: italic;
-    fontWeight: bold;
-    color: #4b4949;
-    textAlign: center;
-    fontSize: 22px;
+
+.login-box-right {
+    background: #EBFAF4;
+    height: 765px;
+    top: 0px;
+    width: 65%;
+}
+
+.login-photo {
+    height: 765px;
+    top: 0px;
 }
