--- src/assets/components/login/Login.js
@@ -1,41 +1,0 @@
-import React, { Component, useState } from 'react';
-import 'firebase/auth';
-import { useFirebaseApp, useUser } from 'reactfire';
-import './Login.css';
-
-
-export default (props) => {
-    const [email, setEmail] = useState('');
-    const [password, setPassword] = useState('');
-
-    const firebase = useFirebaseApp();
-    const user = useUser();
-
-    const submit = async () => {
-        await firebase.auth().createUserWithEmailAndPassword(email, password);
-    }
-
-    const login = async () => {
-        await firebase.auth().signInWithEmailAndPassword(email, password);
-    }
-
-    const logout = async () => {
-        await firebase.auth().signOut();
-    }
-
-    return (
-        <div>
-            
-            {user && <button onClick={logout}>Cerrar sesión</button>}
-            {
-                !user &&
-                <div>
-                    <input type="email" id="email" placeholder="correo electrónico" className="inputLogin" onChange={(event) => setEmail(event.target.value)} />
-                    <input type="password" id="password" placeholder="contraseña" className="inputLogin" onChange={(event) => setPassword(event.target.value)} />
-                    <button onClick={submit} className="btnLogin">Crear cuenta</button>
-                    <button onClick={login} className="btnLogin">Iniciar sesión</button>
-                </div>
-            }
-        </div>
-    )
-}
